5/14提交

+++ bupt-hotel-2/src/main/java/com/example/service/impl/UserServiceImpl.java
@@ -3,20 +3,17 @@ package com.example.service.impl;
 import com.example.common.SystemParam;
 import com.example.dao.entity.AirConditionerDO;
 import com.example.dao.entity.DetailedFeesDO;
-import com.example.dao.entity.RunningQueueDO;
 import com.example.dao.entity.WaitingQueueDO;
 import com.example.dao.mapper.AirConditionerMapper;
 import com.example.dao.mapper.DetailedFeesMapper;
 import com.example.dao.mapper.RunningQueueMapper;
 import com.example.dao.mapper.WaitingQueueMapper;
+import com.example.dto.UserUpdateDTO;
 import com.example.service.UserService;
+import com.example.utils.CommonTool;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
 
 @Service
 @Slf4j
@@ -27,35 +24,31 @@ public class UserServiceImpl implements UserService {
     private final WaitingQueueMapper waitingQueueMapper;
     private final DetailedFeesMapper detailedFeesMapper;
     private final RunningQueueMapper runningQueueMapper;
-    private final SystemParam systemParam = SystemParam.getInstance();
 
     @Override
     public void openOrClose(String acNumber) {
-        AirConditionerDO airConditionerDO = airConditionerMapper.get(acNumber);
-        boolean isOpening = airConditionerDO.getOpening();
-        boolean isConnecting = airConditionerDO.getConnecting();
+        AirConditionerDO airConditionerDO = airConditionerMapper.select(acNumber);
+        boolean workable = airConditionerDO.isWorkable();
+        boolean isOpening = airConditionerDO.isOpening();
+        boolean isConnecting = airConditionerDO.isConnecting();
 
-        AirConditionerDO newAirConditionerDO = AirConditionerDO.builder()
-                .acNumber(acNumber)
-                .opening(!isOpening)
-                .temperature(systemParam.getDefaultTemperature())
-                .windSpeed(systemParam.getDefaultWindSpeed())
-                .build();
+        if (!workable)
+            throw new RuntimeException();
         // 如果空调开着：需要关闭，计算费用，更新空调信息，释放资源
         if (isOpening) {
             if (isConnecting) {
-                log.info("关闭空调" + acNumber + "，释放空调连接");
-                DetailedFeesDO detailedFeesDO = getDetailedFeesDO(airConditionerDO);
+                log.info("关闭空调{}，释放空调连接", acNumber);
+                DetailedFeesDO detailedFeesDO = CommonTool.getDetailedFeesDO(airConditionerDO);
                 detailedFeesMapper.insert(detailedFeesDO);
                 // 释放资源，更新空调信息，系统当前连接数 -1
-                systemParam.setCurrConnectionCount(systemParam.getCurrConnectionCount() - 1);
+                SystemParam.CURR_CONNECTION_COUNT -= 1;
                 runningQueueMapper.delete(acNumber);
 
-                newAirConditionerDO.setConnecting(false);
-                newAirConditionerDO.setCurrFee(airConditionerDO.getCurrFee() + detailedFeesDO.getFee());
+                airConditionerDO.setConnecting(false);
+                airConditionerDO.setCurrFee(airConditionerDO.getCurrFee() + detailedFeesDO.getFee());
             } else {
+                log.info("关闭空调{}，从等待队列中移出", acNumber);
                 waitingQueueMapper.delete(acNumber);
-                log.info("关闭空调" + acNumber + "，从等待队列中移出");
             }
         }
         // 如果空调关着：需要打开，请求连接，更新空调信息
@@ -63,127 +56,54 @@ public class UserServiceImpl implements UserService {
             log.info("打开空调");
             waitingQueueMapper.insert(WaitingQueueDO.builder()
                     .acNumber(acNumber)
-                    .temperature(systemParam.getDefaultTemperature())
-                    .windSpeed(systemParam.getDefaultWindSpeed())
+                    .temperature(SystemParam.DEFAULT_TEMPERATURE)
+                    .windSpeed(SystemParam.DEFAULT_WIND_SPEED)
                     .requestTime(System.currentTimeMillis())
                     .build());
         }
-        airConditionerMapper.update(newAirConditionerDO);
+
+        airConditionerDO.setOpening(!airConditionerDO.isOpening());
+        airConditionerDO.setTemperature(SystemParam.DEFAULT_TEMPERATURE);
+        airConditionerDO.setWindSpeed(SystemParam.DEFAULT_WIND_SPEED);
+        airConditionerMapper.update(airConditionerDO);
     }
 
     /**
-     * 更新温度。
-     *
-     * @param acNumber          空调编号
-     * @param targetTemperature 目标温度
+     * 用户调温调风
+     * @param userUpdateDTO 用户发送的更新请求类
      */
     @Override
-    public void updateTemperature(String acNumber, Integer targetTemperature) {
-        AirConditionerDO airConditionerDO = airConditionerMapper.get(acNumber);
-        // 如果空调opening字段 = false
-        if (!airConditionerDO.getOpening()) {
-            return;
+    public void update(UserUpdateDTO userUpdateDTO) {
+        AirConditionerDO airConditionerDO = airConditionerMapper.select(userUpdateDTO.getAcNumber());
+        boolean workable = airConditionerDO.isWorkable();
+        boolean isOpening = airConditionerDO.isOpening();
+        boolean isConnecting = airConditionerDO.isConnecting();
+
+        if (!workable)
+            throw new RuntimeException();
+        // 如果空调是关闭的，提示用户先打开空调
+        if (!isOpening) {
+            throw new RuntimeException();
         }
         long currentTime = System.currentTimeMillis();
-        if (airConditionerDO.getConnecting()) {
-            DetailedFeesDO detailedFeesDO = getDetailedFeesDO(airConditionerDO);
+        if (isConnecting) {
+            log.info("空调{}正在出风，更新温度为{}，风速为{}", userUpdateDTO.getAcNumber(), userUpdateDTO.getTargetTemperature(), userUpdateDTO.getTargetWindSpeed());
+            DetailedFeesDO detailedFeesDO = CommonTool.getDetailedFeesDO(airConditionerDO);
             detailedFeesMapper.insert(detailedFeesDO);
-            airConditionerMapper.update(AirConditionerDO.builder()
-                    .acNumber(acNumber)
-                    .temperature(targetTemperature)
-                    // 调温调风时，也需要更新
-                    .connectionTime(currentTime)
-                    .currFee(airConditionerDO.getCurrFee() + detailedFeesDO.getFee())
-                    .build());
-            log.info("空调" + acNumber + "连接中，更新温度" + targetTemperature);
+            // 更新空调信息
+            airConditionerDO.setTemperature(userUpdateDTO.getTargetTemperature());
+            airConditionerDO.setWindSpeed(userUpdateDTO.getTargetWindSpeed());
+            airConditionerDO.setCurrFee(airConditionerDO.getCurrFee() + detailedFeesDO.getFee());
         } else {
-            WaitingQueueDO newWaitingQueueDO = WaitingQueueDO.builder()
-                    .acNumber(acNumber)
-                    .windSpeed(systemParam.getDefaultWindSpeed())
-                    .temperature(targetTemperature)
-                    .requestTime(currentTime)
-                    .build();
-            WaitingQueueDO waitingQueueDO = waitingQueueMapper.select(acNumber);
-            newWaitingQueueDO.setWindSpeed(waitingQueueDO.getWindSpeed());
-            newWaitingQueueDO.setRequestTime(waitingQueueDO.getRequestTime());
-            waitingQueueMapper.delete(acNumber);
-            waitingQueueMapper.insert(newWaitingQueueDO);
-            log.info("空调" + acNumber + "在等待队列中，更新温度" + targetTemperature);
+            log.info("空调{}没有出风，更新等待队列中的记录", userUpdateDTO.getAcNumber());
+            WaitingQueueDO waitingQueueDO = waitingQueueMapper.select(userUpdateDTO.getAcNumber());
+            waitingQueueMapper.delete(userUpdateDTO.getAcNumber());
+
+            waitingQueueDO.setTemperature(userUpdateDTO.getTargetTemperature());
+            waitingQueueDO.setWindSpeed(userUpdateDTO.getTargetWindSpeed());
+            waitingQueueDO.setRequestTime(currentTime);
+            waitingQueueMapper.insert(waitingQueueDO);
         }
-        List<RunningQueueDO> runningQueue1 = runningQueueMapper.getAll();
-        System.out.println(runningQueue1);
-        List<WaitingQueueDO> waitQueue1 = waitingQueueMapper.getAll();
-        System.out.println(waitQueue1);
     }
 
-    /**
-     * 更新风速。
-     *
-     * @param acNumber        空调编号
-     * @param targetWindSpeed 目标风速
-     */
-    @Override
-    public void updateWindSpeed(String acNumber, Integer targetWindSpeed) {
-        AirConditionerDO airConditionerDO = airConditionerMapper.get(acNumber);
-        // 如果空调opening字段 = false
-        if (!airConditionerDO.getOpening()) {
-            return;
-        }
-        long currentTime = System.currentTimeMillis();
-        if (airConditionerDO.getConnecting()) {
-            DetailedFeesDO detailedFeesDO = getDetailedFeesDO(airConditionerDO);
-            detailedFeesMapper.insert(detailedFeesDO);
-            airConditionerMapper.update(AirConditionerDO.builder()
-                    .acNumber(acNumber)
-                    .windSpeed(targetWindSpeed)
-                    // 调温调风时，也需要更新
-                    .connectionTime(currentTime)
-                    .currFee(airConditionerDO.getCurrFee() + detailedFeesDO.getFee())
-                    .build());
-            log.info("空调" + acNumber + "连接中，更新风速" + targetWindSpeed);
-        } else {
-            WaitingQueueDO newWaitingQueueDO = WaitingQueueDO.builder()
-                    .acNumber(acNumber)
-                    .temperature(systemParam.getDefaultTemperature())
-                    .windSpeed(targetWindSpeed)
-                    .requestTime(currentTime)
-                    .build();
-            WaitingQueueDO waitingQueueDO = waitingQueueMapper.select(acNumber);
-            newWaitingQueueDO.setWindSpeed(waitingQueueDO.getWindSpeed());
-            newWaitingQueueDO.setRequestTime(waitingQueueDO.getRequestTime());
-            waitingQueueMapper.delete(acNumber);
-            waitingQueueMapper.insert(newWaitingQueueDO);
-            log.info("空调" + acNumber + "在等待队列中，更新风速" + targetWindSpeed);
-        }
-        List<RunningQueueDO> runningQueue1 = runningQueueMapper.getAll();
-        System.out.println(runningQueue1);
-        List<WaitingQueueDO> waitQueue1 = waitingQueueMapper.getAll();
-        System.out.println(waitQueue1);
-    }
-
-    /**
-     * 一个专门计算费用的方法
-     * @param airConditionerDO 空调信息
-     * @return 详细费用对象
-     */
-    private DetailedFeesDO getDetailedFeesDO(AirConditionerDO airConditionerDO) {
-        long currTime = System.currentTimeMillis();
-        long lastConnectionTime = airConditionerDO.getConnectionTime();
-        // TODO 这里也是 按秒算钱
-        long minutes = (lastConnectionTime + currTime) / (1000);
-        int lastTemperature = airConditionerDO.getTemperature();
-        int lastWindSpeed = airConditionerDO.getWindSpeed();
-        // 计算费用
-        LinkedList<HashMap<Integer, Double>> priceTable = systemParam.getPriceTable();
-        Double pricePerMinute = priceTable.get(lastWindSpeed).get(lastTemperature);
-        Double fee = pricePerMinute * minutes;
-        return DetailedFeesDO.builder()
-                .acNumber(airConditionerDO.getAcNumber())
-                .windSpeed(lastWindSpeed)
-                .temperature(lastTemperature)
-                .startTime(lastConnectionTime)
-                .minutes(minutes)
-                .fee(fee)
-                .build();
-    }
 }
